test: adds test environment w/ inputs for dextr
