+++ test/inference.test.js
@@ -1,67 +1,94 @@
 import '@babel/polyfill'
 import * as tf from '@tensorflow/tfjs';
+import npyjs from 'npyjs';
 
 import indexedDbService from '../src/index';
 import load from '../src/load';
 import utils from '../src/utils/utils';
-import iris from './mocks/iris';
+import iris from './mocks/data/iris';
+import Upsampling from './mocks/Upsampling';
 
 const IRIS_MODEL = 'https://storage.googleapis.com/tfjs-models/tfjs/iris_v1/model.json';
+const DEXTR_MODEL = 'http://localhost:5000/static/models/dextr/model.json';
 
 describe('inference', async () => {
-  beforeEach(async function() {
-    await utils.deleteDatabase();
-    await tf.disposeVariables();
+  describe('IRIS MODEL', async () => {
+    beforeEach(async function() {
+      await utils.deleteDatabase();
+      await tf.disposeVariables();
+    });
+  
+    afterEach(async function() {
+      await utils.deleteDatabase();
+      await tf.disposeVariables();
+    }); 
+  
+    it('does inference with server model', async () => {
+      const testingData = tf.tensor2d(iris.map(item => [
+        item.sepal_length, item.sepal_width, item.petal_length, item.petal_width,
+      ]))
+  
+      const model = await tf.loadLayersModel(IRIS_MODEL);
+      const pred = model.predict(testingData)
+  
+      model.dispose();
+      pred.dispose();
+      testingData.dispose();
+    });
+  
+    it('does same inference between indexedDB and server model', async () => {
+      const testingData = tf.tensor2d(iris.map(item => [
+        item.sepal_length, item.sepal_width, item.petal_length, item.petal_width,
+      ]));
+  
+      await indexedDbService.loadAndStoreLayersModel(IRIS_MODEL, 'iris');
+      const idbModelArtifacts = await indexedDbService.loadFromIndexedDb('iris');
+      const idbModel = await load.convertModelArtifactsToModel(idbModelArtifacts);
+      
+      const model = await tf.loadLayersModel(IRIS_MODEL);
+  
+      const pred = await model.predict(testingData).array();
+      const idbPred = await idbModel.predict(testingData).array();
+  
+      expect(idbPred).toEqual(pred);
+    });
+  
+    it('does same inference between server and preindexedDB (before saving) model', async () => {
+      const testingData = tf.tensor2d(iris.map(item => [
+        item.sepal_length, item.sepal_width, item.petal_length, item.petal_width,
+      ]));
+  
+      const modelArtifacts = await indexedDbService.loadAndStoreLayersModel(IRIS_MODEL, 'iris');
+      const model = await load.convertModelArtifactsToModel(modelArtifacts);
+      
+      const origModel = await tf.loadLayersModel(IRIS_MODEL);
+  
+      const origPred = await origModel.predict(testingData).array();
+      const pred = await model.predict(testingData).array();
+  
+      expect(pred).toEqual(origPred);
+    });
   });
 
-  afterEach(async function() {
-    await utils.deleteDatabase();
-    await tf.disposeVariables();
-  }); 
+  describe('DEXTR MODEL', async () => {
+    beforeEach(async function() {
+      await utils.deleteDatabase();
+      await tf.disposeVariables();
+    });
+  
+    afterEach(async function() {
+      await utils.deleteDatabase();
+      await tf.disposeVariables();
+    }); 
+  
+    it('does inference with server model', async () => {
+      const inputRes = await fetch('./base/test/mocks/data/dextrInput.json');
+      const inputJson = await inputRes.json();
+      const inputTensor = await tf.tensor(inputJson.input);
 
-  it('does inference with original IrisModel', async () => {
-    const testingData = tf.tensor2d(iris.map(item => [
-      item.sepal_length, item.sepal_width, item.petal_length, item.petal_width,
-    ]))
-
-    const model = await tf.loadLayersModel(IRIS_MODEL);
-    const pred = model.predict(testingData)
-
-    model.dispose();
-    pred.dispose();
-    testingData.dispose();
-  });
-
-  it('does same inference between idbModel and original', async () => {
-    const testingData = tf.tensor2d(iris.map(item => [
-      item.sepal_length, item.sepal_width, item.petal_length, item.petal_width,
-    ]));
-
-    await indexedDbService.loadAndStoreLayersModel(IRIS_MODEL, 'iris');
-    const idbModelArtifacts = await indexedDbService.loadFromIndexedDb('iris');
-    const idbModel = await load.convertModelArtifactsToModel(idbModelArtifacts);
-    
-    const model = await tf.loadLayersModel(IRIS_MODEL);
-
-    const pred = await model.predict(testingData).array();
-    const idbPred = await idbModel.predict(testingData).array();
-
-    expect(idbPred).toEqual(pred);
-  });
-
-  it('does same inference between original and converted by us', async () => {
-    const testingData = tf.tensor2d(iris.map(item => [
-      item.sepal_length, item.sepal_width, item.petal_length, item.petal_width,
-    ]));
-
-    const modelArtifacts = await indexedDbService.loadAndStoreLayersModel(IRIS_MODEL, 'iris');
-    const model = await load.convertModelArtifactsToModel(modelArtifacts);
-    
-    const origModel = await tf.loadLayersModel(IRIS_MODEL);
-
-    const origPred = await origModel.predict(testingData).array();
-    const pred = await model.predict(testingData).array();
-
-    expect(pred).toEqual(origPred);
+      tf.serialization.registerClass(Upsampling);
+      const model = await tf.loadLayersModel(DEXTR_MODEL);
+      const pred = model.predict(inputTensor)
+    });
   });
 });
