--- test/mocks/iris.js
@@ -1,25 +1,0 @@
-const iris = [
-  {
-    "sepal_length": 5.4,
-    "sepal_width": 3.9,
-    "petal_length": 1.7,
-    "petal_width": 0.4,
-    "species": "setosa"
-  },
-  {
-    "sepal_length": 5.9,
-    "sepal_width": 3,
-    "petal_length": 5.1,
-    "petal_width": 1.8,
-    "species": "virginica"
-  },
-  {
-    "sepal_length": 5.7,
-    "sepal_width": 2.9,
-    "petal_length": 4.2,
-    "petal_width": 1.3,
-    "species": "versicolor"
-  }
-]
-
-export default iris;
